// src/app/api/api/public-catalog-controller.service.ts
export interface ListProductsRequestParams {
    page?: number;
    size?: number;
    sort?: string;
    category?: Array<string>;
    minPrice?: number;
    maxPrice?: number;
    q?: string;
    onSale?: boolean;
    inStock?: boolean;
}